requirements.txt
fastapi==0.76.0
uvicorn==0.17.6
pyJWT==2.3.0
python-decouple==3.6
email-validator==1.2.1
SQLAlchemy==1.4.36
psycopg2-binary==2.9.3
redis==4.3.1